+++ src/components/bulk-email-tool/bulk-email-form/test/BulkEmailForm.test.jsx
@@ -10,11 +10,15 @@ import {
 import BulkEmailForm from '..';
 import * as bulkEmailFormApi from '../data/api';
 import { BulkEmailContext, BulkEmailProvider } from '../../bulk-email-context';
+import { formatDate } from '../../../../utils/formatDateAndTime';
 
 jest.mock('../../text-editor/TextEditor');
 
 const appendMock = jest.spyOn(FormData.prototype, 'append');
 const dispatchMock = jest.fn();
+
+const tomorrow = new Date();
+tomorrow.setDate(new Date().getDate() + 1);
 
 function renderBulkEmailForm() {
   return (
@@ -103,7 +107,7 @@ describe('bulk-email-form', () => {
     fireEvent.click(submitButton);
     const continueButton = await screen.findByRole('button', { name: /continue/i });
     fireEvent.click(continueButton);
-    expect(screen.getByText('Date and time cannot be blank'));
+    expect(screen.getByText('Date and time cannot be blank, and must be a date in the future'));
   });
   test('Adds scheduling data to POST requests when schedule is selected', async () => {
     const postBulkEmailInstructorTask = jest.spyOn(bulkEmailFormApi, 'postBulkEmailInstructorTask');
@@ -116,12 +120,12 @@ describe('bulk-email-form', () => {
     const submitButton = screen.getByText('Schedule Email');
     const scheduleDate = screen.getByTestId('scheduleDate');
     const scheduleTime = screen.getByTestId('scheduleTime');
-    fireEvent.change(scheduleDate, { target: { value: '2020-05-24' } });
+    fireEvent.change(scheduleDate, { target: { value: formatDate(tomorrow) } });
     fireEvent.change(scheduleTime, { target: { value: '10:00' } });
     fireEvent.click(submitButton);
     const continueButton = await screen.findByRole('button', { name: /continue/i });
     fireEvent.click(continueButton);
-    expect(appendMock).toHaveBeenCalledWith('schedule', expect.stringContaining('2020-05-24'));
+    expect(appendMock).toHaveBeenCalledWith('schedule', expect.stringContaining(formatDate(tomorrow)));
     expect(postBulkEmailInstructorTask).toHaveBeenCalledWith(expect.any(FormData), expect.stringContaining('test'));
   });
   test('will PATCH instead of POST when in edit mode', async () => {
@@ -134,7 +138,7 @@ describe('bulk-email-form', () => {
           emailBody: 'test',
           emailSubject: 'test',
           emailRecipients: ['test'],
-          scheduleDate: '2020-05-24',
+          scheduleDate: formatDate(tomorrow),
           scheduleTime: '10:00',
           schedulingId: 1,
           emailId: 1,
